refactor(test-optimization): move Jest test utilities
