+++ packages/dd-trace/src/plugins/util/jest.js
@@ -1,10 +1,6 @@
 'use strict'
 
-const { readFileSync } = require('fs')
-const { parse } = require('../../../vendor/dist/jest-docblock')
-
-const { getTestSuitePath } = require('../../dd-trace/src/plugins/util/test')
-const log = require('../../dd-trace/src/log')
+const { getTestSuitePath, isMarkedAsUnskippable } = require('./test')
 
 /**
  * There are two ways to call `test.each` in `jest`:
@@ -67,62 +63,6 @@ function getRawJestTestName (test) {
 
 function getJestTestName (test) {
   return removeSeedSuffixFromTestName(getRawJestTestName(test))
-}
-
-const globalDocblockRegExp = /^\s*(\/\*\*?(.|\r?\n)*?\*\/)/
-const MAX_COMMENTS_CHECKED = 10
-
-function isMarkedAsUnskippable (test) {
-  let testSource
-
-  try {
-    testSource = readFileSync(test.path, 'utf8')
-  } catch {
-    return false
-  }
-
-  const re = globalDocblockRegExp
-  re.lastIndex = 0
-  let commentsChecked = 0
-
-  while (testSource.length) {
-    const match = re.exec(testSource)
-    if (!match) break
-    const comment = match[1]
-
-    let docblocks
-    try {
-      docblocks = parse(comment)
-    } catch {
-      // Skip unparsable comment and continue scanning
-      if (commentsChecked++ >= MAX_COMMENTS_CHECKED) {
-        return false
-      }
-      continue
-    }
-
-    if (docblocks?.datadog) {
-      try {
-        // @ts-expect-error The datadog type is defined by us and may only be a string.
-        return JSON.parse(docblocks.datadog).unskippable
-      } catch {
-        // If the @datadog block comment is present but malformed, we'll run the suite
-        log.warn('@datadog block comment is malformed.')
-        return true
-      }
-    }
-
-    if (commentsChecked++ >= MAX_COMMENTS_CHECKED) {
-      return false
-    }
-
-    // To stop as soon as no doc blocks are found, slice the source. That way the
-    // regexp works by using the `^` anchor. Without it, it would continue
-    // scanning the rest of the file.
-    testSource = testSource.slice(match[0].length)
-  }
-
-  return false
 }
 
 function getJestSuitesToRun (skippableSuites, originalTests, rootDir, fallbackRootDir) {
@@ -195,6 +135,5 @@ module.exports = {
   getJestTestName,
   getRawJestTestName,
   getJestSuitesToRun,
-  isMarkedAsUnskippable,
   removeSeedSuffixFromTestName,
 }
