+++ packages/dd-trace/test/plugins/util/fixtures/jest/test-to-skip.js
@@ -1,8 +1,8 @@
 'use strict'
 
-// Fixture for util.spec.js. `getJestSuitesToRun` reads this file to scan for
+// Fixture for jest.spec.js. `getJestSuitesToRun` reads this file to scan for
 // a `@datadog` docblock; the body of the suite is never executed.
 
-describe('test-to-run', () => {
+describe('test-to-skip', () => {
   it('is a placeholder fixture', () => {})
 })
