+++ packages/dd-trace/test/plugins/util/fixtures/jest/test-unskippable.js
@@ -1,9 +1,10 @@
+/* eslint-disable jsdoc/valid-types -- @datadog accepts a JSON value rather than a JSDoc type. */
 /**
  * @datadog {"unskippable": true}
  */
 'use strict'
 
-// Fixture for util.spec.js. `getJestSuitesToRun` reads this file to scan for
+// Fixture for jest.spec.js. `getJestSuitesToRun` reads this file to scan for
 // the `@datadog` docblock above; the body of the suite is never executed.
 
 describe('test-unskippable', () => {
